+++ src/utils/i18n/translations/ru.js
@@ -142,6 +142,7 @@ export const ru = {
 
     language: {
       interfaceLanguage: 'Язык интерфейса',
+      showQuickAccessEdge: 'Отображать край шторки',
     },
 
     storage: {